Rename modules with builtin names to prevent confusion
