+++ twtxt/cli.py
@@ -16,14 +16,14 @@ import sys
 import click
 
 from twtxt.config import Config
-from twtxt.file import get_local_tweets, add_local_tweet
+from twtxt.tweetfile import get_local_tweets, add_local_tweet
 from twtxt.helper import run_post_tweet_hook
 from twtxt.helper import style_tweet, style_source, style_source_with_status
 from twtxt.helper import validate_created_at, validate_text
 from twtxt.helper import sort_and_truncate_tweets
-from twtxt.http import get_remote_tweets, get_remote_status
+from twtxt.twhttp import get_remote_tweets, get_remote_status
 from twtxt.log import init_logging
-from twtxt.types import Tweet, Source
+from twtxt.models import Tweet, Source
 
 logger = logging.getLogger(__name__)
